Implement batch Dataset writing with replace and insert modes
Also, when a `stopOnError` option is set to `true`, the batches will be
written one after another, thus when an error occurs, the operation will
stop on the last unsuccessful batch (default mode).
When this option is set to `false`, all scheduled batches will be written
asynchronously in parallel, and the possible errors will be returned only
when all scheduled operations are finished.