Move global mapper variable to the task closure context level
Highly loaded and concurrent Spark environment produced some strange
OptionalDataException errors when deserializing the write tasks.
That appeared to be a problem with an object that is changed concurrently
while being serialized, thus resulting in corrupted serialized data.

Using some deduction and exclusion, it became clear that the object with
such race condition is the mapper, which is apparently changed by
different threads using the same RDD for different Tarantool spaces.

With this patch it is not more possible to use a custom mapper, but the
proper implementation of that feature needs now fixing the mapper race
condition possibility in the driver and re-implementing passing the
mapper without implicits.